Fix loops with permalink:
Explanations of important changes:
- decoding of URL permalink must be done first
  - URL listener moved from map to app
  - => the listener is ready before any map service can modify the state
- decoding of permalink must be done only once
  - unsuscribe from routerNavEnd$ when the URL has been read
